+++ storage/src/mmr/bitmap.rs
@@ -76,7 +76,7 @@ const PRUNED_CHUNKS_PREFIX: u8 = 1;
 
 impl<H: CHasher, const N: usize> MerkleizedBitMap<H, N> {
     /// The size of a chunk in bits.
-    pub const CHUNK_SIZE_BITS: u64 = BitMap::<N>::CHUNK_SIZE_BITS;
+    pub const CHUNK_SIZE_BITS: usize = BitMap::<N>::CHUNK_SIZE_BITS;
 
     /// Return a new empty bitmap.
     pub fn new() -> Self {
@@ -198,7 +198,7 @@ impl<H: CHasher, const N: usize> MerkleizedBitMap<H, N> {
 
     /// Return the number of bits currently stored in the bitmap, irrespective of any pruning.
     #[inline]
-    pub fn len(&self) -> u64 {
+    pub fn len(&self) -> usize {
         self.bitmap.len()
     }
 
@@ -210,7 +210,7 @@ impl<H: CHasher, const N: usize> MerkleizedBitMap<H, N> {
 
     /// Return the number of bits that have been pruned from this bitmap.
     #[inline]
-    pub fn pruned_bits(&self) -> u64 {
+    pub fn pruned_bits(&self) -> usize {
         self.bitmap.pruned_bits()
     }
 
@@ -221,7 +221,7 @@ impl<H: CHasher, const N: usize> MerkleizedBitMap<H, N> {
     /// - Panics if the referenced bit is greater than the number of bits in the bitmap.
     ///
     /// - Panics if there are unprocessed updates.
-    pub fn prune_to_bit(&mut self, bit_offset: u64) {
+    pub fn prune_to_bit(&mut self, bit_offset: usize) {
         let chunk_num = BitMap::<N>::raw_chunk_index(bit_offset);
         if chunk_num < self.bitmap.pruned_chunks() {
             return;
@@ -242,7 +242,7 @@ impl<H: CHasher, const N: usize> MerkleizedBitMap<H, N> {
     /// Return the last chunk of the bitmap and its size in bits. The size can be 0 (meaning the
     /// last chunk is empty).
     #[inline]
-    pub fn last_chunk(&self) -> (&[u8; N], u64) {
+    pub fn last_chunk(&self) -> (&[u8; N], usize) {
         self.bitmap.last_chunk()
     }
 
@@ -252,7 +252,7 @@ impl<H: CHasher, const N: usize> MerkleizedBitMap<H, N> {
     ///
     /// Panics if the bit doesn't exist or has been pruned.
     #[inline]
-    pub fn get_chunk(&self, bit_offset: u64) -> &[u8; N] {
+    pub fn get_chunk(&self, bit_offset: usize) -> &[u8; N] {
         self.bitmap.get_chunk(bit_offset)
     }
 
@@ -278,7 +278,7 @@ impl<H: CHasher, const N: usize> MerkleizedBitMap<H, N> {
 
     /// Convert a bit offset into the position of the Merkle tree leaf it belongs to.
     #[inline]
-    pub(crate) fn leaf_pos(bit_offset: u64) -> u64 {
+    pub(crate) fn leaf_pos(bit_offset: usize) -> u64 {
         let chunk_num = BitMap::<N>::raw_chunk_index(bit_offset);
         leaf_num_to_pos(chunk_num as u64)
     }
@@ -289,13 +289,13 @@ impl<H: CHasher, const N: usize> MerkleizedBitMap<H, N> {
     ///
     /// Panics if the bit doesn't exist or has been pruned.
     #[inline]
-    pub fn get_bit(&self, bit_offset: u64) -> bool {
+    pub fn get_bit(&self, bit_offset: usize) -> bool {
         self.bitmap.get_bit(bit_offset)
     }
 
     #[inline]
     /// Get the value of a bit from its chunk.
-    pub fn get_bit_from_chunk(chunk: &[u8; N], bit_offset: u64) -> bool {
+    pub fn get_bit_from_chunk(chunk: &[u8; N], bit_offset: usize) -> bool {
         BitMap::<N>::get_bit_from_chunk(chunk, bit_offset)
     }
 
@@ -304,7 +304,7 @@ impl<H: CHasher, const N: usize> MerkleizedBitMap<H, N> {
     /// # Warning
     ///
     /// The update will not impact the root until `sync` is called.
-    pub fn set_bit(&mut self, bit_offset: u64, bit: bool) {
+    pub fn set_bit(&mut self, bit_offset: usize, bit: bool) {
         // Apply the change to the inner bitmap
         self.bitmap.set_bit(bit_offset, bit);
 
@@ -391,7 +391,7 @@ impl<H: CHasher, const N: usize> MerkleizedBitMap<H, N> {
         Ok(Self::partial_chunk_root(
             hasher.inner(),
             &mmr_root,
-            next_bit,
+            next_bit as u64,
             &last_chunk_digest,
         ))
     }
@@ -405,7 +405,7 @@ impl<H: CHasher, const N: usize> MerkleizedBitMap<H, N> {
         last_chunk_digest: &H::Digest,
     ) -> H::Digest {
         assert!(next_bit > 0);
-        assert!(next_bit < Self::CHUNK_SIZE_BITS);
+        assert!(next_bit < Self::CHUNK_SIZE_BITS as u64);
         hasher.update(mmr_root);
         hasher.update(&next_bit.to_be_bytes());
         hasher.update(last_chunk_digest);
@@ -423,14 +423,14 @@ impl<H: CHasher, const N: usize> MerkleizedBitMap<H, N> {
         hasher: &mut impl Hasher<H>,
         bit_offset: u64,
     ) -> Result<(Proof<H::Digest>, [u8; N]), Error> {
-        assert!(bit_offset < self.len(), "out of bounds");
+        assert!(bit_offset < self.len() as u64, "out of bounds");
         assert!(
             !self.is_dirty(),
             "cannot compute proof with unprocessed updates"
         );
 
-        let leaf_pos = Self::leaf_pos(bit_offset);
-        let chunk = self.get_chunk(bit_offset);
+        let leaf_pos = Self::leaf_pos(bit_offset as usize);
+        let chunk = self.get_chunk(bit_offset as usize);
 
         let (last_chunk, next_bit) = self.bitmap.last_chunk();
         if leaf_pos == self.mmr.size() {
@@ -439,7 +439,7 @@ impl<H: CHasher, const N: usize> MerkleizedBitMap<H, N> {
             // required in the proof: the mmr's root.
             return Ok((
                 Proof {
-                    size: self.len(),
+                    size: self.len() as u64,
                     digests: vec![self.mmr.root(hasher)],
                 },
                 *chunk,
@@ -447,7 +447,7 @@ impl<H: CHasher, const N: usize> MerkleizedBitMap<H, N> {
         }
 
         let mut proof = Proof::<H::Digest>::range_proof(&self.mmr, leaf_pos, leaf_pos).await?;
-        proof.size = self.len();
+        proof.size = self.len() as u64;
         if next_bit == 0 {
             // Bitmap is chunk aligned.
             return Ok((proof, *chunk));
@@ -475,14 +475,14 @@ impl<H: CHasher, const N: usize> MerkleizedBitMap<H, N> {
             debug!(bit_len, bit_offset, "tried to verify non-existent bit");
             return false;
         }
-        let leaf_pos = Self::leaf_pos(bit_offset);
+        let leaf_pos = Self::leaf_pos(bit_offset as usize);
 
         let mut mmr_proof = Proof::<H::Digest> {
-            size: leaf_num_to_pos(bit_len / Self::CHUNK_SIZE_BITS),
+            size: leaf_num_to_pos(bit_len / Self::CHUNK_SIZE_BITS as u64),
             digests: proof.digests.clone(),
         };
 
-        if bit_len % Self::CHUNK_SIZE_BITS == 0 {
+        if bit_len % Self::CHUNK_SIZE_BITS as u64 == 0 {
             return mmr_proof.verify_element_inclusion(hasher, chunk, leaf_pos, root);
         }
 
@@ -504,7 +504,7 @@ impl<H: CHasher, const N: usize> MerkleizedBitMap<H, N> {
                 return false;
             }
             let last_chunk_digest = hasher.digest(chunk);
-            let next_bit = bit_len % Self::CHUNK_SIZE_BITS;
+            let next_bit = bit_len % Self::CHUNK_SIZE_BITS as u64;
             let reconstructed_root = Self::partial_chunk_root(
                 hasher.inner(),
                 &last_digest,
@@ -524,7 +524,7 @@ impl<H: CHasher, const N: usize> MerkleizedBitMap<H, N> {
             }
         };
 
-        let next_bit = bit_len % Self::CHUNK_SIZE_BITS;
+        let next_bit = bit_len % Self::CHUNK_SIZE_BITS as u64;
         let reconstructed_root =
             Self::partial_chunk_root(hasher.inner(), &mmr_root, next_bit, &last_digest);
 
@@ -849,7 +849,7 @@ mod tests {
             }
 
             // Repeat the test after pruning.
-            let start_bit = SHA256_SIZE as u64 * 8 * 2;
+            let start_bit = SHA256_SIZE * 8 * 2;
             bitmap.prune_to_bit(start_bit);
             for bit_pos in (start_bit..bitmap.len()).rev() {
                 let bit = bitmap.get_bit(bit_pos);
@@ -866,7 +866,7 @@ mod tests {
         });
     }
 
-    fn flip_bit<const N: usize>(bit_offset: u64, chunk: &[u8; N]) -> [u8; N] {
+    fn flip_bit<const N: usize>(bit_offset: usize, chunk: &[u8; N]) -> [u8; N] {
         let byte_offset = BitMap::<N>::chunk_byte_offset(bit_offset);
         let mask = BitMap::<N>::chunk_byte_bitmask(bit_offset);
         let mut tmp = chunk.to_vec();
@@ -906,7 +906,7 @@ mod tests {
                 assert_eq!(bitmap.root(&mut hasher).await.unwrap(), root);
                 bitmap.prune_to_bit(prune_to_bit);
                 for i in prune_to_bit..bitmap.len() {
-                    let (proof, chunk) = bitmap.proof(&mut hasher, i).await.unwrap();
+                    let (proof, chunk) = bitmap.proof(&mut hasher, i as u64).await.unwrap();
 
                     // Proof should verify for the original chunk containing the bit.
                     assert!(
@@ -914,7 +914,7 @@ mod tests {
                             &mut hasher,
                             &proof,
                             &chunk,
-                            i,
+                            i as u64,
                             &root
                         ),
                         "failed to prove bit {i}",
@@ -927,7 +927,7 @@ mod tests {
                             &mut hasher,
                             &proof,
                             &corrupted,
-                            i,
+                            i as u64,
                             &root
                         ),
                         "proving bit {i} after flipping should have failed",
@@ -972,9 +972,7 @@ mod tests {
 
             // prune 10 chunks at a time and make sure replay will restore the bitmap every time.
             for i in (10..=FULL_CHUNK_COUNT).step_by(10) {
-                bitmap.prune_to_bit(
-                    i as u64 * MerkleizedBitMap::<Sha256, SHA256_SIZE>::CHUNK_SIZE_BITS,
-                );
+                bitmap.prune_to_bit(i * MerkleizedBitMap::<Sha256, SHA256_SIZE>::CHUNK_SIZE_BITS);
                 bitmap
                     .write_pruned(context.clone(), PARTITION)
                     .await
@@ -993,7 +991,7 @@ mod tests {
                     bitmap.push_chunk_unchecked(&test_chunk(format!("test{j}").as_bytes()));
                 }
                 assert_eq!(bitmap.bitmap.pruned_chunks(), i);
-                assert_eq!(bitmap.len(), FULL_CHUNK_COUNT as u64 * 256);
+                assert_eq!(bitmap.len(), FULL_CHUNK_COUNT * 256);
                 bitmap.sync(&mut hasher).await.unwrap();
                 assert_eq!(bitmap.root(&mut hasher).await.unwrap(), chunk_aligned_root);
